pyproject.toml
[project]
name = "mcp-guardian-news"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.2",
    "httpx>=0.25.0",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
